CFBD service to shared
